Decouple logger with analytics
