+++ wxm-ios/Toolkit/Toolkit/Analytics/FirebaseAnalytics.swift
@@ -12,17 +12,6 @@ import FirebaseAnalytics
 
 struct FirebaseAnalytics: AnalyticsProviderImplementation {
 
-	func logNetworkError(_ networkError: any NetworkError) {
-		let nsError = NSError(domain: networkDomain,
-							  code: networkError.code,
-							  userInfo: networkError.userInfo)
-		logError(nsError)
-	}
-
-	func logError(_ nsError: NSError) {
-		Crashlytics.crashlytics().record(error: nsError)
-	}
-
 	func trackScreen(_ screen: Screen, parameters: [Parameter : ParameterValue]?) {
 		var params: [String: Any] = [AnalyticsParameterScreenName: screen.rawValue]
 		if let additionalParams = parameters?.toEventParamsDictionary {
