+++ wxm-ios/Toolkit/Toolkit/Logger/Logger.swift
@@ -1,0 +1,29 @@
+//
+//  Logger.swift
+//  Toolkit
+//
+//  Created by Pantelis Giazitsis on 13/5/24.
+//
+
+import Foundation
+import FirebaseCrashlytics
+
+public class Logger {
+	public static let shared = Logger()
+
+	private init() {}
+}
+
+// MARK: - Errors
+public extension Logger {
+	func logNetworkError(_ networkError: any NetworkError) {
+		let nsError = NSError(domain: networkDomain,
+							  code: networkError.code,
+							  userInfo: networkError.userInfo)
+		logError(nsError)
+	}
+
+	func logError(_ nsError: NSError) {
+		Crashlytics.crashlytics().record(error: nsError)
+	}
+}
